+++ skgarden/quantile/ensemble.py
@@ -9,9 +9,8 @@ from sklearn.utils import check_random_state
 from sklearn.utils import check_X_y
 from joblib import delayed, Parallel
 
-from skgarden.utils.helpers import _joblib_parallel_args
-
 from ..forest import ForestRegressor
+from ..helpers import _joblib_parallel_args
 
 from .tree import DecisionTreeQuantileRegressor
 from .tree import ExtraTreeQuantileRegressor
